Split adapter.h/cc into reader.h/writer.h. Draft LargeRecordBatch type
