+++ cpp/src/arrow/allocator-test.cc
@@ -15,9 +15,9 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "gtest/gtest.h"
 #include "arrow/allocator.h"
 #include "arrow/test-util.h"
+#include "gtest/gtest.h"
 
 namespace arrow {
