+++ cpp/src/arrow/ipc/metadata.cc
@@ -358,8 +358,8 @@ const RecordBatchMetadata& DictionaryBatchMetadata::record_batch() const {
 // ----------------------------------------------------------------------
 // Conveniences
 
-Status ReadMessage(int64_t offset, int32_t metadata_length,
-    io::RandomAccessFile* file, std::shared_ptr<Message>* message) {
+Status ReadMessage(int64_t offset, int32_t metadata_length, io::RandomAccessFile* file,
+    std::shared_ptr<Message>* message) {
   std::shared_ptr<Buffer> buffer;
   RETURN_NOT_OK(file->ReadAt(offset, metadata_length, &buffer));
